Implemented confirmation dialog when trying to delete post

+++ src/components/PostActionBar.tsx
@@ -1,8 +1,20 @@
-import { Flex, Button, useColorModeValue } from "@chakra-ui/react";
-import React from "react";
-import { useDeletePostMutation, useMeQuery } from "../generated/graphql";
-import { BiComment, BiEditAlt, BiTrash } from "react-icons/bi";
+import {
+  AlertDialog,
+  AlertDialogBody,
+  AlertDialogCloseButton,
+  AlertDialogContent,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogOverlay,
+  Button,
+  Flex,
+  useColorModeValue,
+  useDisclosure,
+} from "@chakra-ui/react";
 import { useRouter } from "next/router";
+import React, { useRef } from "react";
+import { BiComment, BiEditAlt, BiTrash } from "react-icons/bi";
+import { useDeletePostMutation, useMeQuery } from "../generated/graphql";
 
 interface PostActionBarProps {
   postId: number;
@@ -19,6 +31,8 @@ export const PostActionBar: React.FC<PostActionBarProps> = ({
   const [_, deletePost] = useDeletePostMutation();
   const router = useRouter();
   const textColor = useColorModeValue("gray.500", "gray.400");
+  const { isOpen, onOpen, onClose } = useDisclosure();
+  const cancelRef = useRef(null);
 
   const buttonStyle = {
     variant: "ghost",
@@ -28,11 +42,7 @@ export const PostActionBar: React.FC<PostActionBarProps> = ({
   };
 
   return (
-    <Flex
-      onClick={(e) => {
-        e.stopPropagation();
-      }}
-    >
+    <Flex>
       <Button leftIcon={<BiComment />} {...buttonStyle}>
         {comments} Comments
       </Button>
@@ -41,16 +51,41 @@ export const PostActionBar: React.FC<PostActionBarProps> = ({
           <Button leftIcon={<BiEditAlt />} {...buttonStyle}>
             Edit
           </Button>
-          <Button
-            leftIcon={<BiTrash />}
-            {...buttonStyle}
-            onClick={async () => {
-              await deletePost({ id: postId });
-              router.replace("/");
-            }}
-          >
+          <Button leftIcon={<BiTrash />} onClick={onOpen} {...buttonStyle}>
             Delete
           </Button>
+          <AlertDialog
+            motionPreset="slideInBottom"
+            leastDestructiveRef={cancelRef}
+            onClose={onClose}
+            isOpen={isOpen}
+            isCentered
+          >
+            <AlertDialogOverlay />
+
+            <AlertDialogContent>
+              <AlertDialogHeader>Delete post?</AlertDialogHeader>
+              <AlertDialogCloseButton />
+              <AlertDialogBody>
+                Are you sure you want to delete your post?
+              </AlertDialogBody>
+              <AlertDialogFooter>
+                <Button ref={cancelRef} onClick={onClose}>
+                  Cancel
+                </Button>
+                <Button
+                  colorScheme="red"
+                  ml={3}
+                  onClick={async () => {
+                    await deletePost({ id: postId });
+                    router.replace("/");
+                  }}
+                >
+                  Delete post
+                </Button>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </>
       ) : null}
     </Flex>
